# Solve SA6002 in pkg/scraper/client/resource/client.go