@serathius 2023-03-29
Please run the benchmark to confirm no changes in allocs.
You can consult https://github.com/kubernetes-sigs/metrics-server/blob/master/Makefile#L112-L137 on how to compare performance using benchstat