@colega 2023-03-29
Oh sure, these are the benchmark results:
```
goos: linux
goarch: amd64
pkg: sigs.k8s.io/metrics-server/pkg/scraper/client/resource
cpu: 11th Gen Intel(R) Core(TM) i7-11700K @ 3.60GHz
                            │     old     │                new                 │
                            │   sec/op    │   sec/op     vs base               │
KubeletClient_GetMetrics-16   260.2µ ± 2%   257.1µ ± 1%  -1.17% (p=0.015 n=10)

                            │     old      │              new               │
                            │     B/op     │     B/op      vs base          │
KubeletClient_GetMetrics-16   189.8Ki ± 0%   189.9Ki ± 0%  ~ (p=0.971 n=10)

                            │    old     │                new                │
                            │ allocs/op  │ allocs/op   vs base               │
KubeletClient_GetMetrics-16   794.0 ± 0%   793.0 ± 0%  -0.13% (p=0.000 n=10)
```